+++ 9월 10일 - 스택, 큐, 덱/1935.cpp
@@ -3,92 +3,64 @@
 //
 
 #include <iostream>
-#include <algorithm>
 #include <string>
-#include <deque>
+#include <stack>
 
 using namespace std;
 
-int main() {
-    int t;
-    cin >> t;
+int alphabet[26]; //알파벳
 
-    string func;
-    int n;
-    string arr;
-    deque<int> dq;
+int main(void)
+{
+    int N;
+    cin >> N;
+    string str;
+    cin >> str;
 
-    while (t--)
+    //입력 알파벳 배열에 저장 ex) 1이면 alphabet[0]
+    for (int i = 0; i < N; i++)
+        cin >> alphabet[i];
+
+    stack<double> s;
+    for (int i = 0; i < str.length(); i++)
     {
-        cin >> func;
-        cin >> n;
-        cin >> arr;
-
-        //두자리수를 고려해서 string으로 받음
-        string temp;
-        for(int i=0; i<arr.length(); i++){
-            if(arr[i] == '[')
-                continue;
-            else if ('0' <= arr[i] && arr[i] <='9')
-                temp += arr[i];
-            else if (arr[i] == ',' || arr[i] == ']') {
-                if (!temp.empty()) {
-                    dq.push_back(stoi(temp));
-                    temp.clear();
-                }
-            }
-        }
-
-        bool isReversed = false; //뒤집혔는지 검사
-        bool isError= false; //에러인지 검사
-
-        for(int i=0; i<func.length(); i++)
+        //연산자일 경우
+        if (str[i] == '+' || str[i] == '-' || str[i] == '*' || str[i] == '/')
         {
-            //뒤집기 두번 쓰일 수 있어서 *true로 만드는게 아니라 있던걸 반대로 해야함..ㅡㅡ
-            if (func[i]=='R')
-                isReversed = !isReversed;
-            else //'D'일때
-            {
-                if (dq.empty()) //비어있으면 뺄 게 없어서 에러
-                {
-                    isError = true;
-                    cout << "error\n";
-                    break;
-                }
-                else //비어있지 않고
-                    if(isReversed) //뒤집혀있으면
-                        dq.pop_back(); //뒤에서 뺌
-                    else // 그대로면
-                        dq.pop_front(); //앞에서 뺌
-            }
-        }
+            //피연산자 두 개를 빼줌
+            double a = s.top();
+            s.pop();
+            double b = s.top();
+            s.pop();
 
-        //에러가 안났고
-        if (!isError) {
-            //뒤집혀있지 않으면
-            if (!isReversed) {
-                cout << "[";
-                while (!dq.empty()) {
-                    cout << dq.front(); //앞에서 빼서 출력
-                    dq.pop_front();
-                    if (!dq.empty())
-                        cout << ",";
-                }
-                cout << "]\n";
-            }
-            //뒤집혀 있으면
-            else
+            //연산자 종류에 따라 계산해줌. 이때 먼저 뺀 a가 사실은 식 상으로는 뒤에 있는 것 주의
+            switch (str[i])
             {
-                cout << "[";
-                while (!dq.empty()) {
-                    cout << dq.back(); //뒤에서 빼서 출력
-                    dq.pop_back();
-                    if (!dq.empty())
-                        cout << ",";
-                }
-                cout << "]\n";
+                case '+':
+                    s.push(b + a);
+                    break;
+                case '-':
+                    s.push(b - a);
+                    break;
+                case '*':
+                    s.push(b * a);
+                    break;
+                case '/':
+                    s.push(b / a);
+                    break;
             }
         }
+        //피연산자 일 경우 문자열을 int형으로 바꿔서 넣어줌
+        else
+            s.push(alphabet[str[i] - 'A']);
     }
+
+    double result = s.top();
+    //소숫점 2자리 고정
+    cout << fixed;
+    cout.precision(2);
+    cout << result << "\n";
+
     return 0;
+
 }
